requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
deepgram-sdk==2.11.0
distro==1.9.0
Django==5.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.16.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
multidict==6.4.3
numpy==2.2.5
openai==1.76.0
pillow==11.2.1
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
reportlab==4.4.0
service-identity==24.2.0
setuptools==79.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.6.0
yarl==1.20.0
zope.interface==7.2
